moving bits to testUtilities
